test_parseが肥大化してたのでtest_convertに分離
